\name{StanHeaders-package}
\alias{StanHeaders-package}
\alias{StanHeaders}
\docType{package}
\title{
  The Stan C++ Library Headers
}
\description{
  The Stan project implements Markov Chain Monte Carlo and optimization for 
  statistical models. This R package provides the Stan header files so that
  they can easily be used by other R packages via the LinkingTo mechanism.
}
\details{
\tabular{ll}{
Package: \tab StanHeaders\cr
Type: \tab Package\cr
Version: \tab 2.6.1\cr
Date: \tab 2015-02-08\cr
License: \tab BSD 3 clause\cr
}
  There are no R functions, only C++ headers. See the rstan package for an
  illustration of how to use the StanHeaders package via the LinkingTo
  mechanism.
}
\author{
  Ben Goodrich for the R package and the Stan Development Team for Stan
  Maintainer: Ben Goodrich <benjamin.goodrich@columbia.edu>
}
\references{
  http://mc-stan.org
}
\keyword{ package }
\examples{
  # No examples because there are no R functions.
}
